feat: update nonce type to uint64 across GemFactory, IntexFactory, and runtime

+++ contracts/precompiles/src/IGemFactory.sol
@@ -19,7 +19,7 @@ interface IGemFactory {
     ///         MUST equal the caller's current on-chain promis op-nonce (fetch via
     ///         `outbe_deriveKeys` + `IPromis.opNonceOf`) and the bound amount is the
     ///         gem's load. Returns the minted Promis amount.
-    function mineGemPromis(uint256 gemId, uint256 nonce, bytes32 mac, uint64 opNonce) external returns (uint256);
+    function mineGemPromis(uint256 gemId, uint64 nonce, bytes32 mac, uint64 opNonce) external returns (uint256);
     function getStatistics() external view returns (uint256 totalGemsIssued, uint256 totalIntexParked);
 
     // --- GemPosition NFT (ERC-721-style, non-transferable; owner = merchant) ---
